Fix TLAS rebuild synchronization
We synchronized TLAS rebuild command with subsequent CmdTraceRays command.
But we also need to sync TLAS rebuild commad with TLAS rebuild on the
next frame.

TLAS rebuild is not synchronized via some frame high level synchronization
(as opposite to COLOR_ATTACHENT_OUTPUT because we have acquire semaphore).
So we do this via barrier.

Note, this can be quite optimal on real implementations, because
added flags might not change what existing synchronization (for
CmdTraceRays) already does for cache setup.